+++ packages/webawesome/src/components/icon/icon.styles.ts
@@ -19,6 +19,8 @@ export default css`
   :host(:not([auto-width])) {
     width: 1.25em;
     height: 1em;
+    min-width: 1.25em; /* <-- this is what Safari respects for intrinsic */
+    min-height: 1em;
   }
 
   /* Auto-width */